Refactored db for smaller files
