--- server/fadzmaq/api/match_data.py
@@ -1,58 +1,0 @@
-my_match = {
-        "match": {
-            "user_id": "6c18ec1127d5a8951265f35c66b60f82",
-            "name": "Joshua",
-            "photo_location": "https://upload.wikimedia.org/wikipedia/commons/thumb/0/04/John_Legend_2019_by_Glenn_Francis.jpg/1280px-John_Legend_2019_by_Glenn_Francis.jpg",
-            "contact_details": {
-                "phone": "0423 161 368",
-                "email": "joshua.smith@gmail.com"
-            },
-            "profile_fields": [
-                {
-                    "id": 7,
-                    "name": "Location",
-                    "display_value": "Duncraig",
-                    "distance": 0,
-                },
-                {
-                    "id": 3,
-                    "name": "About me",
-                    "display_value": "I am a boxer. Lets get in the ring."
-                }
-            ],
-            "hobbies": [
-                {
-                    "share": [
-                        {
-                            "id": 33,
-                            "name": "Boxing"
-                        }
-                    ]
-                },
-                {
-                    "discover": [
-                        {
-                            "id": 99,
-                            "name": "Netball"
-                        }
-                    ]
-                }
-            ]
-        }
-    }
-
-
-my_matches = {
-        "matches": [
-            {
-                "id": "6c18ec1127d5a8951265f35c66b60f82",
-                "name": "Joshua",
-                "photo": "https://upload.wikimedia.org/wikipedia/commons/thumb/0/04/John_Legend_2019_by_Glenn_Francis.jpg/1280px-John_Legend_2019_by_Glenn_Francis.jpg"
-            },
-            {
-                "id": "2ff8557a16f674e466ee4ae619f22758",
-                "name": "Mike",
-                "photo": "https://upload.wikimedia.org/wikipedia/commons/thumb/0/04/John_Legend_2019_by_Glenn_Francis.jpg/1280px-John_Legend_2019_by_Glenn_Francis.jpg"
-            }
-        ]
-    }
