--- server/fadzmaq/api/profile_data.py
@@ -1,47 +1,0 @@
-my_profile = {
-        "profile": {
-            "user_id": "6c18ec1127d5a8951265f35c66b60f82",
-            "name": "Joshua",
-            "age": 20,
-            "birth-date": "28/08/1999",
-            "photo_location": "https://upload.wikimedia.org/wikipedia/commons/thumb/0/04/John_Legend_2019_by_Glenn_Francis.jpg/1280px-John_Legend_2019_by_Glenn_Francis.jpg",
-            "contact_details": {
-                "phone": "0423 161 368",
-                "email": "joshua.smith@gmail.com"
-            },
-            "profile_fields": [
-                {
-                    "id": 2,
-                    "name": "Location",
-                    "display_value": "Duncraig",
-                    "distance": 0,
-                    "lat": 55.7558,
-                    "long": 37.6173,
-                    "time": 1566966101,
-                },
-                {
-                    "id": 1,
-                    "name": "About me",
-                    "display_value": "I am a boxer. Lets get in the ring."
-                }
-            ],
-            "hobbies": [
-                {
-                    "share": [
-                        {
-                            "id": 33,
-                            "name": "Boxing"
-                        }
-                    ]
-                },
-                {
-                    "discover": [
-                        {
-                            "id": 99,
-                            "name": "Netball"
-                        }
-                    ]
-                }
-            ]
-        }
-    }
