+++ server/fadzmaq/api/recs_data.py
@@ -1,183 +1,98 @@
-my_recs = {
-        "results": [
-            {
-                "user": {
-                    "user_id": "f1bd8af2a28adbf70af43cfc6d79ef95",
-                    "name": "Michaela",
-                    "photo_location": "https://upload.wikimedia.org/wikipedia/commons/thumb/0/04/John_Legend_2019_by_Glenn_Francis.jpg/1280px-John_Legend_2019_by_Glenn_Francis.jpg",
-                    "profile_fields": [
-                        {
-                            "id": 2,
-                            "name": "Location",
-                            "display_value": "Sorrento",
-                            "distance": 8
-                        },
-                        {
-                            "id": 3,
-                            "name": "About me",
-                            "display_value": "Like a bit of everything, into netball and scuba diving!"
-                        }
-                    ],
-                    "hobbies": [
-                        {
-                            "share": [
-                                {
-                                    "id": 5,
-                                    "name": "Scuba Diving"
-                                },
-                                {
-                                   "id": 99,
-                                   "name": "Netball"
-                                }
-                            ]
-                        },
-                        {
-                            "discover": [
-                                {
-                                    "id": 5,
-                                    "name": "Scuba Diving"
-                                },
-                                {
-                                    "id": 99,
-                                    "name": "Netball"
-                                },
-                                {
-                                    "id": 20,
-                                    "name": "Road Cycling"
-                                }
-                            ]
-                        }
-                    ]
-                }
-            },
-            {
-                "user": {
-                    "user_id": "6c18ec1127d5a8951265f35c66b60f82",
-                    "name": "Joshua",
-                    "photo_location": "https://upload.wikimedia.org/wikipedia/commons/thumb/0/04/John_Legend_2019_by_Glenn_Francis.jpg/1280px-John_Legend_2019_by_Glenn_Francis.jpg",
-                    "profile_fields": [
-                        {
-                            "id": 7,
-                            "name": "Location",
-                            "display_value": "Duncraig",
-                            "distance": 3
-                        },
-                        {
-                            "id": 3,
-                            "name": "About me",
-                            "display_value": "I am a boxer. Lets get in the ring."
-                        }
-                    ],
-                    "hobbies": [
-                        {
-                            "share": [
-                                {
-                                    "id": 33,
-                                    "name": "Boxing"
-                                }
-                            ]
-                        },
-                        {
-                            "discover": [
-                                {
-                                    "id": 99,
-                                    "name": "Netball"
-                                }
-                            ]
-                        }
-                    ]
-                }
-            },
-            {
-                "user": {
-                    "user_id": "857a7932ce146081c59d2ba659a506f4",
-                    "name": "Jody",
-                    "age": 23,
-                    "gender": 2,
-                    "photo_location": "https://upload.wikimedia.org/wikipedia/commons/thumb/0/04/John_Legend_2019_by_Glenn_Francis.jpg/1280px-John_Legend_2019_by_Glenn_Francis.jpg",
-                    "profile_fields": [
-                        {
-                            "id": 5,
-                            "name": "Location",
-                            "display_value": "Duncraig",
-                            "distance": 3
-                        },
-                        {
-                            "id": 3,
-                            "name": "About me",
-                            "display_value": "Chill and looking for hobby buddies."
-                        }
-                    ],
-                    "hobbies": [
-                        {
-                            "share": [
-                                {
-                                    "id": 12,
-                                    "name": "Rock climbing"
-                                }
-                            ]
-                        },
-                        {
-                            "discover": [
-                                {
-                                    "id": 12,
-                                    "name": "Rock climbing"
-                                }
-                            ]
-                        }
-                    ]
-                }
+recs = {
+    "recommendations": [
+        {
+            "rank": 1,
+            "user": {
+                "user_id": "29f51c08adac957424e06699b81acdb5",
+                "name": "John",
+                "photo_location": "URL",
+                "profile_fields": [
+                    {
+                        "name": "About me",
+                        "display_value": "Avid rock climber and hiking enthusiast."
+                    },
+                    {
+                        "name": "location",
+                        "display_value": "<5"
+                    }
+                ],
+                "hobbies": [
+                     {
+                         "container": "share",
+                         "hobbies": [
+                             {
+                                 "id": 3,
+                                 "name": "Rock Climbing"
+                             }
+                         ]
+                     },
+                     {
+                         "container": "discover",
+                         "hobbies": [
+                             {
+                                 "id": 1,
+                                 "name": "Boxing"
+                             }
+                        ]
+                    },
+                    {
+                        "container": "matched",
+                         "hobbies": [
+                             {
+                                 "id": 1,
+                                 "name": "Boxing"
+                             }
+                        ]
+                    }
+                ]
             }
-        ]
-    }
-
-my_candiate = {
-                "user": {
-                    "user_id": "f1bd8af2a28adbf70af43cfc6d79ef95",
-                    "name": "Michaela",
-                    "photo_location": "https://upload.wikimedia.org/wikipedia/commons/thumb/0/04/John_Legend_2019_by_Glenn_Francis.jpg/1280px-John_Legend_2019_by_Glenn_Francis.jpg",
-                    "profile_fields": [
-                        {
-                            "id": 2,
-                            "name": "Location",
-                            "display_value": "Sorrento",
-                            "distance": 8
-                        },
-                        {
-                            "id": 3,
-                            "name": "About me",
-                            "display_value": "Like a bit of everything, into netball and scuba diving!"
-                        }
-                    ],
-                    "hobbies": [
-                        {
-                            "share": [
-                                {
-                                    "id": 5,
-                                    "name": "Scuba Diving"
-                                },
-                                {
-                                   "id": 99,
-                                   "name": "Netball"
-                                }
-                            ]
-                        },
-                        {
-                            "discover": [
-                                {
-                                    "id": 5,
-                                    "name": "Scuba Diving"
-                                },
-                                {
-                                    "id": 99,
-                                    "name": "Netball"
-                                },
-                                {
-                                    "id": 20,
-                                    "name": "Road Cycling"
-                                }
-                            ]
-                        }
-                    ]
-                }
+        },
+        {
+            "rank": 2,
+            "user": {
+                "user_id": "29f51c08adac957424e06699b81acdb5",
+                "name": "Amy",
+                "photo_location": "URL",
+                "profile_fields": [
+                    {
+                        "name": "About me",
+                        "display_value": "Avid rock climber and hiking enthusiast."
+                    },
+                    {
+                        "name": "location",
+                        "display_value": "6"
+                    }
+                ],
+                "hobbies": [
+                     {
+                         "container": "share",
+                         "hobbies": [
+                             {
+                                 "id": 3,
+                                 "name": "Rock Climbing"
+                             }
+                         ]
+                     },
+                     {
+                         "container": "discover",
+                         "hobbies": [
+                             {
+                                 "id": 1,
+                                 "name": "Boxing"
+                             }
+                        ]
+                    },
+                    {
+                        "container": "matched",
+                         "hobbies": [
+                             {
+                                 "id": 1,
+                                 "name": "Boxing"
+                             }
+                        ]
+                    }
+                ]
             }
+        }
+    ]
+}
