+++ server/fadzmaq/database/__init__.py
@@ -1,0 +1,14 @@
+# @file
+# @brief        Database functions
+# database/__init__.py
+#
+# FadZmaq Project
+# Professional Computing. Semester 2 2019
+#
+# Copyright FadZmaq © 2019      All rights reserved.
+# @author Lachlan Russell       22414249@student.uwa.edu.au
+# @author Jordan Russell        [email]
+
+from flask import current_app, g
+from sqlalchemy import create_engine
+import fadzmaq
