+++ server/fadzmaq/database/connection.py
@@ -1,0 +1,26 @@
+def init_app(app):
+    app.teardown_appcontext(close_db)
+
+
+def get_engine():
+    if fadzmaq.engine is None:
+        print ("new engine")
+        fadzmaq.engine = create_engine(current_app.config['DATABASE_URI'])
+    return fadzmaq.engine
+
+
+def get_db():
+    if 'db' not in g:
+        g.db = connect_db()
+    return g.db
+
+
+def close_db(e=None):
+    db = g.pop('db', None)
+
+    if db is not None:
+        db.close()
+
+
+def connect_db():
+    return get_engine().connect()
