+++ server/fadzmaq/database/recommendations.py
@@ -1,0 +1,1 @@
+import fadzmaq.database as db
